+++ laim/util.py
@@ -1,5 +1,6 @@
 import os
 import pwd
+import re
 
 
 def drop_privileges(user):
@@ -10,3 +11,10 @@ def drop_privileges(user):
 
     os.setgid(pwd_details.pw_gid)
     os.setuid(pwd_details.pw_uid)
+
+
+def unfold(folded):
+    '''Helper to unfold headers'''
+    if folded is None:
+        return None
+    return re.sub(r'\r?\n ', ' ', folded)