+++ tests/integration/test_handler.py
@@ -2,7 +2,8 @@ from unittest import mock
 
 import pytest
 
-from laim import Laim, TaskArguments
+from laim import Laim
+from laim.laim import TaskArguments
 
 pytestmark = pytest.mark.integration
 
@@ -10,10 +11,10 @@ pytestmark = pytest.mark.integration
 def test_handler(temp_config):
     class Handler(Laim):
         def handle_message(self, sender, recipients, message):
-            self.stop_event.set()
+            self.stop()
 
-    with mock.patch('laim.drop_privileges'):
-        with mock.patch('laim.LaimController'):
+    with mock.patch('laim.laim.drop_privileges'):
+        with mock.patch('laim.laim.LaimController'):
             handler = Handler(config_file=temp_config)
 
     handler.queue.put(TaskArguments(mock.Mock(), [], b''))
@@ -25,11 +26,11 @@ def test_handler(temp_config):
 def test_crashing_handler(temp_config):
     class Handler(Laim):
         def handle_message(self, sender, recipients, message):
-            self.stop_event.set()
+            self.stop()
             raise ValueError()
 
-    with mock.patch('laim.drop_privileges'):
-        with mock.patch('laim.LaimController'):
+    with mock.patch('laim.laim.drop_privileges'):
+        with mock.patch('laim.laim.LaimController'):
             handler = Handler(config_file=temp_config)
 
     handler.queue.put(TaskArguments(mock.Mock(), [], b''))