+++ tools/lint.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+set -e
+
+./venv/bin/pylint --rcfile .pylintrc laim --jobs 0 --score=no